Add optimize functionality to dev-tools GUI
Moves core logic of optimization to xdl and imports from within expo-cli and
dev-tools. Adds a new Graph QL mutation which the dev-tools front end
calls to trigger the xdl optimizeAsync function.
