+++ packages/expo-cli/src/commands/optimize.js
@@ -1,16 +1,8 @@
-import glob from 'glob';
-import sharp from 'sharp';
-import path from 'path';
-import chalk from 'chalk';
-import crypto from 'crypto';
-import fs from 'fs';
-import { ProjectUtils } from 'xdl';
-import JsonFile from '@expo/json-file';
+import { Project, ProjectUtils } from 'xdl';
 import prompt from '../prompt';
 import log from '../log';
 
 export async function action(projectDir = './', options = {}) {
-  log(chalk.green('Optimizing assets...'));
   const { exp } = await ProjectUtils.readConfigJsonAsync(projectDir);
   if (exp === null) {
     log.warn('No Expo configuration found. Are you sure this is a project directory?');
@@ -28,157 +20,8 @@ export async function action(projectDir = './', options = {}) {
       options.save = true;
     }
   }
-
-  const [assetJson, assetInfo] = await readAssetJsonAsync(projectDir);
-  // Keep track of which hash values in assets.json are no longer in use
-  const outdated = new Set();
-  for (const fileHash in assetInfo) outdated.add(fileHash);
-
-  // Filter out image assets
-  let totalSaved = 0;
-  const files = getAssetFiles(exp, projectDir);
-  const regex = /\.(png|jpg|jpeg)$/;
-  const images = files.filter(file => regex.test(file.toLowerCase()));
-  for (const image of images) {
-    const hash = calculateHash(image);
-    if (assetInfo[hash]) {
-      outdated.delete(hash);
-      continue;
-    }
-    const { size: prevSize } = fs.statSync(image);
-
-    const newName = createNewFilename(image);
-    await optimizeImage(image, newName);
-
-    const { size: newSize } = fs.statSync(image);
-    const amountSaved = prevSize - newSize;
-    if (amountSaved < 0) {
-      // Delete the optimized version and revert changes
-      fs.renameSync(newName, image);
-      assetInfo[hash] = true;
-      log(
-        chalk.gray(
-          `Compressed version of ${image} was larger than original. Using original instead.`
-        )
-      );
-      continue;
-    }
-    // Recalculate hash since the image has changed
-    const newHash = calculateHash(image);
-    assetInfo[newHash] = true;
-
-    if (options.save) {
-      if (hash === newHash) {
-        log(
-          chalk.gray(
-            `Compressed asset ${image} is identical to the original. Using original instead.`
-          )
-        );
-        fs.unlinkSync(newName);
-      } else {
-        log(chalk.gray(`Saving original asset to ${newName}`));
-        // Save the old hash to prevent reoptimizing
-        assetInfo[hash] = true;
-      }
-    } else {
-      // Delete the renamed original asset
-      fs.unlinkSync(newName);
-    }
-    totalSaved += amountSaved;
-    log.warn(`Saved ${toReadableValue(amountSaved)}`);
-  }
-  if (totalSaved === 0) {
-    log('No assets optimized. Everything is fully compressed!');
-  } else {
-    log(`Finished compressing assets. ${chalk.green(toReadableValue(totalSaved))} saved.`);
-  }
-  // Remove assets that have been deleted/modified from assets.json
-  outdated.forEach(outdatedHash => {
-    delete assetInfo[outdatedHash];
-  });
-  assetJson.writeAsync(assetInfo);
+  await Project.optimizeAsync(projectDir, options);
 }
-
-const toReadableValue = bytes => {
-  const sizes = ['Bytes', 'KB', 'MB', 'GB'];
-  const index = Math.floor(Math.log(bytes) / Math.log(1024));
-  const reduced = (bytes / Math.pow(1024, index)).toFixed(2) * 1;
-
-  return `${reduced} ${sizes[index]}`;
-};
-
-/*
- * Calculate SHA256 Checksum value of a file based on its contents
- */
-const calculateHash = file => {
-  const contents = fs.readFileSync(file);
-  return crypto
-    .createHash('sha256')
-    .update(contents)
-    .digest('hex');
-};
-
-/*
- * Compress an inputted jpg or png and save original copy with .expo extension
- */
-const optimizeImage = async (image, newName) => {
-  log(`Optimizing ${image}`);
-  // Rename the file with .expo extension
-  fs.copyFileSync(image, newName);
-
-  // Extract the format and compress
-  const buffer = await sharp(image).toBuffer();
-  const { format } = await sharp(buffer).metadata();
-  if (format === 'jpeg') {
-    await sharp(newName)
-      .jpeg({ quality: 60 })
-      .toFile(image)
-      .catch(err => log.error(err));
-  } else {
-    await sharp(newName)
-      .png({ quality: 60 })
-      .toFile(image)
-      .catch(err => log.error(err));
-  }
-};
-
-/*
- * Find all project assets under assetBundlePatterns in app.json excluding node_modules
- */
-const getAssetFiles = (exp, projectDir) => {
-  const { assetBundlePatterns } = exp;
-  const options = { cwd: projectDir, ignore: '**/node_modules/**' };
-  const files = [];
-  assetBundlePatterns.forEach(pattern => {
-    files.push(...glob.sync(pattern, options));
-  });
-  return files.map(file => `${projectDir}/${file}`.replace('//', '/'));
-};
-
-/*
- * Read the contents of assets.json under .expo-shared folder. Create the file/directory if they don't exist.
- */
-const readAssetJsonAsync = async projectDir => {
-  const dirPath = path.join(projectDir, '.expo-shared');
-  if (!fs.existsSync(dirPath)) {
-    fs.mkdirSync(dirPath);
-  }
-
-  const assetJson = new JsonFile(path.join(dirPath, 'assets.json'));
-  if (!fs.existsSync(assetJson.file)) {
-    await assetJson.writeAsync({});
-  }
-  const assetInfo = await assetJson.readAsync();
-  return [assetJson, assetInfo];
-};
-
-/*
- * Add .expo extension to a filename in a path string
- */
-const createNewFilename = image => {
-  const { dir, name, ext } = path.parse(image);
-  return dir + '/' + name + '.expo' + ext;
-};
 
 export default program => {
   program
