+++ packages/xdl/package.json
@@ -83,6 +83,7 @@
     "request-promise-native": "1.0.5",
     "resolve-from": "4.0.0",
     "semver": "5.5.0",
+    "sharp": "^0.22.1",
     "slugid": "1.1.0",
     "slugify": "1.3.1",
     "source-map-support": "0.4.18",
